requirements.txt
# ============================================================================
# CE7412 Enzyme Classification — Dependencies
# ============================================================================
# Install: pip install -r requirements.txt
# For PyG: pip install torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-{TORCH_VERSION}.html
# ============================================================================

# Core
torch>=2.1.0
torchvision>=0.16.0

# PyTorch Geometric
torch-geometric>=2.4.0
# torch-scatter and torch-sparse must be installed separately (see above)

# Protein data
proteinshake>=0.3.0
fair-esm>=2.0.0

# ML utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Data & config
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0

# Logging
tqdm>=4.65.0
tensorboard>=2.14.0
wandb>=0.15.0          # optional: set logging.use_wandb=true in config

# Utilities
networkx>=3.1
rich>=13.0.0           # pretty terminal output
